Move visualizer callbacks to their own module
