+++ anomalib/utils/callbacks/__init__.py
@@ -31,8 +31,7 @@ from .min_max_normalization import MinMaxNormalizationCallback
 from .model_loader import LoadModelCallback
 from .tiler_configuration import TilerConfigurationCallback
 from .timer import TimerCallback
-from .visualizer_image import VisualizerCallbackImage
-from .visualizer_metric import VisualizerCallbackMetric
+from .visualizer import VisualizerCallbackImage, VisualizerCallbackMetric
 
 __all__ = [
     "CdfNormalizationCallback",
