+++ anomalib/utils/callbacks/visualizer/visualizer_base.py
