+++ anomalib/utils/callbacks/visualizer/visualizer_image.py
