+++ anomalib/utils/callbacks/visualizer/visualizer_metric.py
